requirements.txt
flask==2.3.3
flask-socketio==5.3.6
cryptography==42.0.5
requests==2.31.0
python-socketio==5.11.2
eventlet==0.33.3
pyOpenSSL==24.1.0
mysql-connector-python==9.2.0

